Add team members sequentially
IN NET_LBFO implementation, the MAC address of the bond net
adapter will be that one of the last member of the net adapters
list. In adaptive-load balancing, alb, tlb or even lacp, the switch
might expect that the bonded net adapter will have the first nic's
MAC.

To address this issue, a bond will be created with the first nic only,
and afterwards the other team members will be added.